* experiment/sdl: (43 commits)
  organization
  Prevent the MSVC compiler from warning about the clang pragma.
  Many minor fixes to address LLVM compiler warnings.
  The BOOL type was defined differently depending on if an objective-c++ or cpp file included the defining header. Including objc.h in std_Platform.h seems to fix this issue.
  Fixes to make the app build without errors with the gfx_OpenGL code. Many of the platform specific macOS code is stubbed out and will need to be fully implemented later.
  The macOS OpenGL code needed to include the library files differently (from the OpenGL framework). Also, I added some shim header entries for the palleted texture stuff. No current video card supports this anymore, so we'll need to implement a new way. For now, this shim helps the build compile.
  Added the gfx_OpenGL sub library source code directly into the nTA macOS project. The Mac app will only need the OpenGL version for now so I didn't bother creating a dylib project for this. This can be changed later if we need to add another gfx implementation (ES, Metal, etc).
  Minor tweaks on Windows for compatibility.
  Many minor fixes to address LLVM compiler warnings.
  Upon further investigation, DWORD *needs* to be defined as a 32 bit unsigned integer; several codec functions use DWORD and WORD to assume 32 bit and 16 values respectively. So DWORD is uint32_t. This causes a problem in the messaging code where pointers are sent around using the DWORD dwParamA, and dwParamB. This is *so* bad; but it's there now and will be hard to route out. For now, I changed the type of all dwParamA, dwParamB to my own custom Param_t type. This is a DWORD on Windows and size_t on macOS. This should keep pointers from being truncated as they are passed around. Yikes.
  Many minor fixes to address LLVM compiler warnings.
  Implemented more macOS specific stubs for fe_Skirmish, fe_MultiplayerSetup, and app_nTA.
  Implemented some of the macOS sys_App methods. sys_App::Create is just a stub for now; it will need to be revisited when the Cocoa Window+App code is worked out.
  Implemented the std_FileIO functions for macOS.
  Link zlib.
  Prevent a MSVC compiler warning on %f with 0 literal.
  Addressed some const'ness in the gfx interface. Any reference parameters that are actually input only should be const references; I believe this was the intent.
  autogenerated
  Map Windows __min and __max to MIN and MAX.
  typedef'd ssize_t and socklen_t on Windows for compatibility.
  ...

+++ nTA/Source/app_nTA_Exposed.inl
@@ -131,7 +131,7 @@ LRESULT CALLBACK app_nTA::ft_AddScanType( LPVOID lpParams )
 		return 1;
 	}
 
-	theFileTree.AddFileType( Ext, Priority );
+	theFileTree.AddFileType( Ext, (int)Priority );
 
 	return 1;
 }
@@ -160,7 +160,7 @@ LRESULT CALLBACK app_nTA::ft_AddSubs( LPVOID lpParams )
 		return 1;
 	}
 
-	theFileTree.AddSubDirectoriesAsLibs( Priority );
+	theFileTree.AddSubDirectoriesAsLibs( (int)Priority );
 
 	return 1;
 }
